+++ gDRplots/tests/testthat/test-plot-sa-response_curve.R
@@ -56,6 +56,7 @@ test_that("plot_dose_response_sa works as expected", {
   expect_equal(sort(ggplot2::get_guide_data(plt_3, "colour")[[".label"]]), sort(sel_grp_names))
   expect_length(unique(ggplot2::ggplot_build(plt_3)$data[[3]]$colour), 1) # curve data only for 1 cell line
   
+  # scenario: lack of metric data at all -> plot only observation
   plt_4 <- plot_dose_response_sa(dt_metrics = dt_metrics[normalization_type == "RV", ],
                                  dt_average = dt_average,
                                  selection_name = selected_drug,
@@ -63,9 +64,10 @@ test_that("plot_dose_response_sa works as expected", {
   expect_is(plt_4, "gg")
   expect_equal(plt_4[["labels"]][["y"]], "GR")
   expect_true(any(grepl("Concentration", plt_4[["labels"]][["x"]])))
-  expect_length(plt_4[["layers"]], 0)
-  expect_length(ggplot2::ggplot_build(plt_4)$data[[1]], 0)
+  expect_length(plt_4[["layers"]], 2)
+  expect_length(ggplot2::ggplot_build(plt_4)[["data"]], 2) # lack of fit lines
   
+  # scenario: plot by drugs 
   group_var <- drug_name
   selected_celline <- "cellline_BA"
   normalization_type <- "RV"
@@ -103,7 +105,7 @@ test_that("plot_dose_response_sa works as expected", {
   expect_equal(NROW(unique(ggplot2::ggplot_build(plt_6)[["data"]][[3]][["colour"]])), 
                NROW(intersect(drug_name_subset, drug_name_vec))) # metric 
   
-  # scenario: lack of metric data for selected `group_names`
+  # scenario: lack of averager data for selected `group_names`
   drug_name_subset <- c("drug_002", "drug_003", "drug_004", "drug_005", "drug_006")
   dt_average_lack <- data.table::copy(dt_average)[get(drug_name) %in% drug_name_subset]
   drug_name_vec <- c("drug_005", "drug_006", "drug_007", "drug_008", "drug_009", "drug_010")
@@ -130,8 +132,8 @@ test_that("plot_dose_response_sa works as expected", {
   expect_is(plt_8, "gg")
   expect_equal(ggplot2::get_guide_data(plt_8, "colour")[[".label"]],
                drug_name_vec) # legend is present without "drug_xx"
-
-  # scenario: NA i concentration column
+  
+  # scenario: NA in concentration column -> only fitted curve
   dt_average_NA <- data.table::copy(dt_average)
   dt_average_NA[["Concentration"]] <- NA
   
@@ -143,8 +145,9 @@ test_that("plot_dose_response_sa works as expected", {
   expect_equal(plt_9[["labels"]][["y"]], "GR")
   expect_true(any(grepl("Concentration", plt_9[["labels"]][["x"]])))
   expect_true(grepl(selected_celline, plt_9[["labels"]][["title"]]))
-  expect_length(ggplot2::ggplot_build(plt_9)$data[[1]], 0) # no data at all
-
+  expect_length(ggplot2::ggplot_build(plt_9)$data[[1]], 7)
+  expect_equal(ggplot2::get_guide_data(plt_9, "colour")[[".label"]],
+               sort(unique(dt_metrics[["DrugName"]])))
   
   # scenario: values for drug are NAs for selected cell line
   dt_metrics_NA <- data.table::copy(dt_metrics)
@@ -166,7 +169,7 @@ test_that("plot_dose_response_sa works as expected", {
   expect_true(grepl(selected_celline, plt_10[["labels"]][["title"]]))
   expect_length(ggplot2::ggplot_build(plt_10)$data[[1]], 0) # no data at al
   
-  # scenario: there is no data in metrics for selected cell line
+  # scenario: there is no data in metrics for selected cell line -> plot only observations
   dt_metrics_miss <- data.table::copy(dt_metrics)
   dt_metrics_miss <- dt_metrics_miss[get(cellline_name) != selected_celline, ]
   
@@ -178,9 +181,9 @@ test_that("plot_dose_response_sa works as expected", {
   
   expect_is(plt_11, "gg")
   expect_true(grepl(selected_celline, plt_11[["labels"]][["title"]]))
-  expect_length(ggplot2::ggplot_build(plt_11)$data[[1]], 0) # no data at all
+  expect_length(ggplot2::ggplot_build(plt_11)$data[[1]], 7) 
   
-  # scenario: there is no data in average for selected cell line
+  # scenario: there is no data in average for selected cell line -> only fiited curve
   dt_average_miss <- data.table::copy(dt_average)
   dt_average_miss <- dt_average_miss[get(cellline_name) != selected_celline, ]
   
@@ -189,10 +192,11 @@ test_that("plot_dose_response_sa works as expected", {
                                   selection_name = selected_celline,
                                   group_var = group_var,
                                   group_names = drug_name_vec)
-  
   expect_is(plt_12, "gg")
   expect_true(grepl(selected_celline, plt_12[["labels"]][["title"]]))
-  expect_length(ggplot2::ggplot_build(plt_12)$data[[1]], 0) # no data at all
+  expect_length(ggplot2::ggplot_build(plt_12)$data[[1]], 7) # no data at all
+  expect_equal(ggplot2::get_guide_data(plt_12, "colour")[[".label"]],
+               drug_name_vec)
   
   # scenario: there is no data in metrics and average for selected cell line
   plt_13 <- plot_dose_response_sa(dt_metrics = dt_metrics_miss,
@@ -203,7 +207,23 @@ test_that("plot_dose_response_sa works as expected", {
   
   expect_is(plt_13, "gg")
   expect_equal(plt_13[["labels"]][["title"]], selected_celline)
-  expect_length(ggplot2::ggplot_build(plt_12)$data[[1]], 0) # no data at all
+  expect_length(ggplot2::ggplot_build(plt_13)$data[[1]], 0) # no data at all
+  
+  # scenario: there is no data in average for selected cell line and maxlog10Concentration is NA
+  dt_metrics_no_maxconc <- data.table::copy(dt_metrics)
+  dt_metrics_no_maxconc[["maxlog10Concentration"]] <- NA
+  plt_14 <- plot_dose_response_sa(dt_metrics = dt_metrics_no_maxconc,
+                                  dt_average = dt_average_miss,
+                                  selection_name = selected_celline,
+                                  group_var = group_var,
+                                  group_names = drug_name_vec)
+  expect_is(plt_14, "gg")
+  expect_true(grepl(selected_celline, plt_14[["labels"]][["title"]]))
+  expect_length(ggplot2::ggplot_build(plt_14)$data[[1]], 7) # no data at all
+  expect_equal(ggplot2::get_guide_data(plt_12, "colour")[[".label"]],
+               drug_name_vec)
+  expect_equal(ggplot2::layer_scales(plt_14)$x$range$range,
+               c(-3.5, 2.0))
   
   # scenario: fitted curve has bigger y-range than avg points
   drug_nm <- "drug_0A"
@@ -240,15 +260,15 @@ test_that("plot_dose_response_sa works as expected", {
                                                              sel_metrics$ec50,
                                                              sel_metrics$h))
   
-  plt_14 <- plot_dose_response_sa(dt_metrics = tab_met, 
+  plt_15 <- plot_dose_response_sa(dt_metrics = tab_met, 
                                   dt_average = tab_avg,
                                   selection_name = "drug_0A",
                                   group_var = cellline_name,
                                   group_names = "cl_14",
                                   normalization_type = "GR")
-  expect_is(plt_14, "gg")
-  expect_true(grepl(drug_nm, plt_14[["labels"]][["title"]]))
-  plot_range <- range(as.numeric(ggplot2::layer_scales(plt_14)$y$get_labels()))
+  expect_is(plt_15, "gg")
+  expect_true(grepl(drug_nm, plt_15[["labels"]][["title"]]))
+  plot_range <- range(as.numeric(ggplot2::layer_scales(plt_15)$y$get_labels()))
   expect_true(all(data.table::between(fitted_range, plot_range[1], plot_range[2])))
   
   # testing assertion